refactor(SortBlock): renamed component file to more correct version
